fix: correct type errors in Meetup API

+++ src/controllers/meetup-controller.ts
@@ -1,71 +1,80 @@
-import { Request, Response, NextFunction } from 'express';
+import {NextFunction, Request, Response} from 'express';
 import meetupService from "../services/meetup-service";
 import ApiError from "../exceptions/api-error";
 import {UserRole} from "../constants/userRoles";
-import {NextFunction} from "express";
-import {AuthenticatedRequest} from "../types/AuthenticatedRequest";
 import {CreateMeetup} from "../types/CreateMeetup";
 import {MeetUpSearchQuery} from "../types/MeetUpSearchQuery";
 import {UpdateMeetup} from "../types/UpdateMeetup";
 import {ErrorMessages} from "../constants/errorMessages";
 
 class MeetupController {
-    async findAllMeetups(req: Request, res: Response, next: NextFunction): Promise<Response>   {
+    async findAllMeetups(req: Request, res: Response, next: NextFunction): Promise<Response | void>   {
         try {
             const meetups = await meetupService.findAllMeetups(req.query as MeetUpSearchQuery);
             return res.status(200).json(meetups);
         } catch (err) {
             next(err);
+            return;
         }
     }
 
-    async findMeetupById(req: Request, res: Response, next: NextFunction): Promise<Response>   {
+    async findMeetupById(req: Request, res: Response, next: NextFunction): Promise<Response | void>   {
         try {
            const { id } = req.params;
            const meetup = await meetupService.findMeetupById(+id);
            return res.status(200).json(meetup)
         } catch (err) {
             next(err);
+            return;
         }
     }
 
-    async addMeetup(req: AuthenticatedRequest, res: Response, next: NextFunction): Promise<Response>   {
+    async addMeetup(req: Request, res: Response, next: NextFunction): Promise<Response | void>   {
         try {
-            const { role, id } = req.user;
-            if (role === UserRole.USER) {
-                throw ApiError.Forbidden(ErrorMessages.USER_FORBIDDEN);
+            if (req.user) {
+                const { role, id } = req.user;
+                if (role === UserRole.USER) {
+                    throw ApiError.Forbidden(ErrorMessages.USER_FORBIDDEN);
+                }
+                const meetup = await meetupService.addMeetup(req.validatedData as CreateMeetup, id);
+                return res.status(201).json(meetup);
             }
-            const meetup = await meetupService.addMeetup(req.validatedData as CreateMeetup, id);
-            return res.status(201).json(meetup);
         } catch (err) {
             next(err);
+            return;
         }
     }
 
-    async updateMeetup(req: AuthenticatedRequest, res: Response, next: NextFunction): Promise<Response>   {
+    async updateMeetup(req: Request, res: Response, next: NextFunction): Promise<Response | void>   {
         try {
-            const { role, id } = req.user;
-            const {userId} = req.validatedData;
-            if (role === UserRole.USER && userId !== id) {
-                throw ApiError.Forbidden(ErrorMessages.USER_FORBIDDEN);
+            if (req.user && 'role' in req.user) {
+                const { role, id } = req.user;
+                const { userId } = req.validatedData as UpdateMeetup;
+                if (role === UserRole.USER && userId !== id) {
+                    throw ApiError.Forbidden(ErrorMessages.USER_FORBIDDEN);
+                }
+                const updatedMeetup = await meetupService.updateMeetup(req.validatedData as UpdateMeetup);
+                return res.status(200).json(updatedMeetup);
             }
-            const updatedMeetup = await meetupService.updateMeetup(req.validatedData as UpdateMeetup);
-            return res.status(200).json(updatedMeetup);
         } catch (err) {
             next(err);
+            return;
         }
     }
 
-    async deleteMeetup(req: AuthenticatedRequest, res: Response, next: NextFunction): Promise<Response>   {
+    async deleteMeetup(req: Request, res: Response, next: NextFunction): Promise<Response | void>   {
         try {
-            const { role, id } = req.user;
-            if (role === UserRole.USER) {
-                throw ApiError.Forbidden(ErrorMessages.USER_FORBIDDEN);
+            if (req['user']) {
+                const { role, id } = req['user'];
+                if (role === UserRole.USER) {
+                    throw ApiError.Forbidden(ErrorMessages.USER_FORBIDDEN);
+                }
+                const deletedMeetup = await meetupService.deleteMeetup(+req.params.id, id);
+                return res.status(204).json(deletedMeetup);
             }
-            const deletedMeetup = await meetupService.deleteMeetup(+req.params.id, id);
-            return res.status(204).json(deletedMeetup);
         } catch (err) {
             next(err);
+            return;
         }
     }
 }
